destroy 짜기

+++ Circle-03/philosophers/philo/include/entities.h
@@ -5,12 +5,12 @@
 
 typedef struct s_data
 {
-	int				value;
+	long long				value;
 	pthread_mutex_t	lock;
 }	t_data;
 
 int		init_data(t_data *data);
-int		sdata(t_data *data, int new_value);
-int		gdata(t_data *data);
+int		sdata(t_data *data, long long new_value);
+long long		gdata(t_data *data);
 
 #endif
